Implement change request for simplifying subscription retrieval code, more default values for fetch Stricter content types, more granular error handling, more entries in the log, some updated tests Custom registry host on all in one sart Update all schemas to not use union types, update publish validation to be in line with the confluent client Update default internal error response, update offset tests to not use publishing , as the actual offsets are only used in case of a join / rebalance Update default params for consumer creation, update test headers for compatibility with test registry Update test headers to be in line with rest proxy, refactor partition id validation code Remote tests pass except base64 / json validation on publish Scrap json validation, in line with proxy, leave b64 one as is Add todo list, revert container file changes Split rest code, drop TODO list ,to be recreated as issues Move error message function to parent class, drop unused validation code Add rest apis main file
