+++ karapace/kafka_rest_apis/__main__.py
@@ -1,0 +1,29 @@
+from karapace import version as karapace_version
+from karapace.kafka_rest_apis import KafkaRest
+
+import argparse
+import os
+import sys
+
+
+def main():
+    parser = argparse.ArgumentParser(prog="karapace rest", description="Karapace: Your Kafka essentials in one tool")
+    parser.add_argument("--version", action="version", help="show program version", version=karapace_version.__version__)
+    parser.add_argument("config_file", help="configuration file path")
+    arg = parser.parse_args()
+
+    if not os.path.exists(arg.config_file):
+        print("Config file: {} does not exist, exiting".format(arg.config_file))
+        return 1
+
+    kc = KafkaRest(arg.config_file)
+    try:
+        return kc.run(host=kc.config["host"], port=kc.config["port"])
+    except Exception:  # pylint: disable-broad-except
+        if kc.raven_client:
+            kc.raven_client.captureException(tags={"where": "karapace_rest"})
+        raise
+
+
+if __name__ == "__main__":
+    sys.exit(main())
