+++ karapace/kafka_rest_apis/admin.py
@@ -1,0 +1,133 @@
+from kafka import KafkaAdminClient
+from kafka.admin import ConfigResource, ConfigResourceType, NewTopic
+from kafka.errors import for_code, UnknownTopicOrPartitionError, UnrecognizedBrokerVersion
+from kafka.future import Future
+from kafka.protocol.admin import DescribeConfigsRequest
+from kafka.protocol.metadata import MetadataRequest
+from kafka.protocol.offset import OffsetRequest, OffsetResetStrategy
+from typing import List
+
+import six
+
+
+class KafkaRestAdminClient(KafkaAdminClient):
+    def get_topic_config(self, topic: str) -> dict:
+        config_version = self._matching_api_version(DescribeConfigsRequest)
+        req_cfgs = [ConfigResource(ConfigResourceType.TOPIC, topic)]
+        cfgs = self.describe_configs(req_cfgs)
+        assert len(cfgs) == 1
+        assert len(cfgs[0].resources) == 1
+        err, _, _, _, config_values = cfgs[0].resources[0]
+        if err != 0:
+            raise for_code(err)
+        topic_config = {}
+        for cv in config_values:
+            if config_version == 0:
+                name, val, _, _, _ = cv
+            else:
+                name, val, _, _, _, _ = cv
+            topic_config[name] = val
+        return topic_config
+
+    def new_topic(self, name: str):
+        self.create_topics([NewTopic(name, 1, 1)])
+
+    def cluster_metadata(self, topics: List[str] = None) -> dict:
+        """List all kafka topics."""
+        resp = {"topics": {}}
+        brokers = set()
+        metadata_version = self._matching_api_version(MetadataRequest)
+        if 1 <= metadata_version <= 6:
+            if not topics:
+                request = MetadataRequest[metadata_version]()
+            else:
+                request = MetadataRequest[metadata_version](topics=topics)
+            future = self._send_request_to_node(self._client.least_loaded_node(), request)
+            self._wait_for_futures([future])
+            response = future.value
+            resp_brokers = response.brokers
+            for b in resp_brokers:
+                if metadata_version == 0:
+                    node_id, _, _ = b
+                else:
+                    node_id, _, _, _ = b
+                brokers.add(node_id)
+            resp["brokers"] = list(brokers)
+            if not response.topics:
+                return resp
+
+            for tup in response.topics:
+                if response.API_KEY != 0:
+                    err, topic, _, partitions = tup
+                else:
+                    err, topic, partitions = tup
+                if err:
+                    raise for_code(err)
+                topic_data = []
+                for part in partitions:
+                    if metadata_version <= 4:
+                        _, partition_index, leader_id, replica_nodes, isr_nodes = part
+                    else:
+                        _, partition_index, leader_id, replica_nodes, isr_nodes, _ = part
+                    isr_nodes = set(isr_nodes)
+                    topic_response = {"partition": partition_index, "leader": leader_id, "replicas": []}
+                    for node in replica_nodes:
+                        topic_response["replicas"].append({
+                            "broker": node,
+                            "leader": node == leader_id,
+                            "in_sync": node in isr_nodes
+                        })
+                    topic_data.append(topic_response)
+                resp["topics"][topic] = {"partitions": topic_data}
+            return resp
+        raise UnrecognizedBrokerVersion(
+            "Kafka Admin interface cannot determine the controller using MetadataRequest_v{}.".format(metadata_version)
+        )
+
+    def make_offsets_request(self, topic: str, partition_id: int, timestamp: int) -> Future:
+        v = self._matching_api_version(OffsetRequest)
+        if v == 0:
+            request = OffsetRequest[0](-1, list(six.iteritems({topic: [(partition_id, timestamp, 1)]})))
+        elif v == 1:
+            request = OffsetRequest[1](-1, list(six.iteritems({topic: [(partition_id, timestamp)]})))
+        else:
+            request = OffsetRequest[2](-1, 1, list(six.iteritems({topic: [(partition_id, timestamp)]})))
+
+        future = self._send_request_to_node(self._client.least_loaded_node(), request)
+        return future
+
+    def get_offsets(self, topic: str, partition_id: int) -> dict:
+        beginning_f = self.make_offsets_request(topic, partition_id, OffsetResetStrategy.EARLIEST)
+        end_f = self.make_offsets_request(topic, partition_id, OffsetResetStrategy.LATEST)
+        self._wait_for_futures([beginning_f, end_f])
+        beginning_resp = beginning_f.value
+        end_resp = end_f.value
+        v = self._matching_api_version(OffsetRequest)
+        assert len(beginning_resp.topics) == 1
+        assert len(end_resp.topics) == 1
+        _, beginning_partitions = beginning_resp.topics[0]
+        _, end_partitions = end_resp.topics[0]
+
+        assert len(beginning_partitions) == 1
+        assert len(end_partitions) == 1
+        if v == 0:
+            assert len(beginning_partitions[0][2]) == 1
+            assert partition_id == beginning_partitions[0][0]
+            assert partition_id == end_partitions[0][0]
+            start_err = beginning_partitions[0][1]
+            end_err = beginning_partitions[0][1]
+            for e in [start_err, end_err]:
+                if e != 0:
+                    raise for_code(e)
+            rv = {
+                "beginning_offset": beginning_partitions[0][2][0],
+                "end_offset": end_partitions[0][2][0],
+            }
+        else:
+            rv = {
+                "beginning_offset": beginning_partitions[0][3],
+                "end_offset": end_partitions[0][3],
+            }
+        if any(val < 0 for val in rv.values()):
+            raise UnknownTopicOrPartitionError("Invalid values for offsets found")
+        return rv
